Remove support for statusbar
As of Sublime text 3.2 there is now some git support in the editor. That means
among other things that there is also a Git icon in the statusbar. Combined with
the fact that the author prefers to have a clean statusbar, the support for git
status in the statusbar has now been removed