Fix bug where commit would not show for some directories
This commit fixes a bug where a non git dir would be added to the list.
If that directory was a parent of a git dir not yet present in the
git_dir list, we would always return false. The bug arose due to a
misunderstanding of the 'in' keyword in python. It has been replaced
with the equality operator, so now we only do exact matches, not
substring matches.